+++ pkg/vrct/vrctFs/tests/fs_test.go
@@ -53,7 +53,7 @@ func TestCreatingFile(t *testing.T) {
 	_ = os.RemoveAll(tmpPath)
 }
 
-func makeFsChanges(t *testing.T, fsVrct *vrctFs.FsVRCT, testFilePath string) {
+func makeFsChanges(t *testing.T, fsVrct *vrctFs.VRCTFs, testFilePath string) {
 	err := fsVrct.CreateFile(testFilePath, []byte(newContent), nil, false, vrctFs.TextFile)
 	if err != nil {
 		t.Fatal("Failed to create file "+testFilePath+"\n", err)
@@ -92,7 +92,7 @@ func makeFsChanges(t *testing.T, fsVrct *vrctFs.FsVRCT, testFilePath string) {
 	}
 }
 
-func revertFsChanges(t *testing.T, fsVrct *vrctFs.FsVRCT, testFilePath string) {
+func revertFsChanges(t *testing.T, fsVrct *vrctFs.VRCTFs, testFilePath string) {
 	err := fsVrct.Revert()
 	if err != nil {
 		t.Fatal("Failed to revert VRCT\n", err)
